Split stats.ex & add a debounce on the manual garbage collection
