--- neurow/lib/metric_plug_exporter.ex
@@ -1,3 +1,0 @@
-defmodule MetricsPlugExporter do
-  use Prometheus.PlugExporter
-end
