+++ neurow/lib/neurow/internal_api/endpoint.ex
@@ -6,7 +6,7 @@ defmodule Neurow.InternalApi.Endpoint do
   alias Neurow.Broker.Message
 
   use Plug.Router
-  plug(MetricsPlugExporter)
+  plug(Neurow.Observability.MetricsPlugExporter)
 
   plug(Neurow.JwtAuthPlug,
     credential_headers: ["x-interservice-authorization", "authorization"],
@@ -16,7 +16,7 @@ defmodule Neurow.InternalApi.Endpoint do
       &Neurow.Configuration.internal_api_verbose_authentication_errors/0,
     max_lifetime: &Neurow.Configuration.internal_api_jwt_max_lifetime/0,
     send_forbidden: &Neurow.InternalApi.Endpoint.send_forbidden/3,
-    inc_error_callback: &Neurow.Stats.Security.inc_jwt_errors_internal/0,
+    inc_error_callback: &Neurow.Observability.SecurityStats.inc_jwt_errors_internal/0,
     exclude_path_prefixes: [
       "/ping",
       "/nodes",
@@ -67,7 +67,10 @@ defmodule Neurow.InternalApi.Endpoint do
     |> put_resp_header("content-type", "application/json")
     |> send_resp(
       200,
-      Jason.encode!(Neurow.Stats.Processes.grouping_stats() |> Enum.map(&Map.from_struct/1))
+      Jason.encode!(
+        Neurow.Observability.SystemStats.process_groups()
+        |> Enum.map(&Map.from_struct/1)
+      )
     )
   end
 
@@ -113,7 +116,7 @@ defmodule Neurow.InternalApi.Endpoint do
           end)
         end)
 
-        Neurow.Stats.MessageBroker.inc_message_published(issuer)
+        Neurow.Observability.MessageBrokerStats.inc_message_published(issuer)
 
         conn
         |> put_resp_header("content-type", "application/json")
