+++ neurow/lib/neurow/observability/metric_plug_exporter.ex
@@ -1,0 +1,3 @@
+defmodule Neurow.Observability.MetricsPlugExporter do
+  use Prometheus.PlugExporter
+end
